feature: payment celery

+++ payments/schemas.py
@@ -1,5 +1,6 @@
 from datetime import date
 from decimal import Decimal
+from pydantic import Field
 
 from ninja import Schema, ModelSchema
 
@@ -32,6 +33,6 @@ class PaymentSlipSchema(Schema):
     email: str
     account_number: str
     total_work_hours: int
-    total_pay: Decimal
+    total_pay: float 
     # start_date: date
     # end_date: date
